Handle compression in membuffer size
The previous method using getStorageSize() only allocated membuffer
based on the size of the stored data, which breaks file reading in the
case where the datasets are compressed. Using the total number of
elements multiplied by the size of the type (char) fixes this.

Also, removed a bunch of logging statements that are no longer
necessary.